refactor(anima_train): 抽 bootstrap / observability / model_loading / text_encoding / state 到 training/
ADR 0003 PR-A 第 1 步：把 anima_train.py 里 5 个无外部依赖的 util 簇按职责
搬到 runtime/training/ 子包：

- training/bootstrap.py        ← deps / yaml / 进度条 (5 fn)
- training/observability.py    ← WandBMonitor + loss 曲线 (4 entry)
- training/model_loading.py    ← prefix 推断 / safetensors / 路径解析 / xformers (9 fn)
- training/text_encoding.py    ← Qwen + T5 加权 tokenize (4 fn)
- training/state.py            ← save/load_training_state

顶层 anima_train.py 用 re-export 保留 sister script / test 对
`anima_train.X` 的访问路径不变。顺手删除死代码 `_lazy_imports()`（grep 0
caller）。

anima_train.py 行数 2901 → 2163；migration test 全绿。
