+++ runtime/anima_train.py
@@ -53,562 +53,38 @@ logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(
 logger = logging.getLogger(__name__)
 
 
-# ============================================================================
-# 依赖检测
-# ============================================================================
-
-def ensure_dependencies(auto_install=False):
-    """检测并可选自动安装缺失依赖"""
-    required = {
-        "numpy": "numpy",
-        "PIL": "Pillow",
-        "safetensors": "safetensors",
-        "transformers": "transformers",
-        "einops": "einops",
-        "torchvision": "torchvision",
-        "yaml": "pyyaml",
-    }
-    missing = []
-    for module_name, pip_name in required.items():
-        try:
-            __import__(module_name)
-        except Exception:
-            missing.append(pip_name)
-    if not missing:
-        return
-    missing_list = ", ".join(sorted(set(missing)))
-    print(f"Missing dependencies: {missing_list}")
-    if not auto_install:
-        print(f"Install them with:\n  {sys.executable} -m pip install {missing_list}")
-        raise SystemExit(1)
-    cmd = [sys.executable, "-m", "pip", "install", *sorted(set(missing))]
-    print("Installing missing dependencies...")
-    try:
-        subprocess.run(cmd, check=False)
-    except Exception as exc:
-        print(f"Auto-install failed: {exc}")
-        raise SystemExit(1)
-    # Re-check after install
-    still_missing = []
-    for module_name, pip_name in required.items():
-        try:
-            __import__(module_name)
-        except Exception:
-            still_missing.append(pip_name)
-    if still_missing:
-        still_list = ", ".join(sorted(set(still_missing)))
-        print(f"Still missing: {still_list}")
-        raise SystemExit(1)
-
-
-# ============================================================================
-# YAML 配置加载
-# ============================================================================
-
-def load_yaml_config(config_path):
-    """加载 YAML 配置文件"""
-    try:
-        import yaml
-    except ImportError:
-        print("PyYAML not installed. Install with: pip install pyyaml")
-        raise SystemExit(1)
-
-    config_path = Path(config_path)
-    if not config_path.exists():
-        raise FileNotFoundError(f"Config file not found: {config_path}")
-
-    with open(config_path, "r", encoding="utf-8") as f:
-        config = yaml.safe_load(f)
-
-    if config is None:
-        config = {}
-
-    return config
-
-
-def apply_yaml_config(args, config):
-    """将 YAML 配置应用到 args；命令行显式参数优先于 YAML。
-
-    实现走 studio.argparse_bridge.merge_yaml_into_namespace —— 字段名 / 默认值
-    都从 studio.schema.TrainingConfig 这一份单一权威源派生，避免与 parse_args
-    脱节。未在 schema 中的 YAML 键会被忽略（拼写错误一目了然）。
-
-    在 merge 前调用 migrate_legacy_attention 兜底老 yaml 的 xformers/flash_attn
-    双 bool —— argparse_bridge 不走 pydantic validator，schema 层的迁移逻辑
-    无法生效，所以这里显式做一次。
-    """
-    from studio.argparse_bridge import merge_yaml_into_namespace
-    from studio.schema import TrainingConfig, migrate_legacy_attention
-    config = migrate_legacy_attention(dict(config or {}))
-    return merge_yaml_into_namespace(args, config, TrainingConfig)
-
-
-# Lazy imports after dependency check
-def _lazy_imports():
-    global np, Image
-    import numpy as np
-    from PIL import Image
-
-
-# ============================================================================
-# 进度和 Loss 曲线可视化
-# ============================================================================
-
-def init_progress(show_progress, total_steps):
-    """初始化 Rich 进度条"""
-    if not show_progress:
-        return None, None, None
-    try:
-        from rich.progress import (
-            BarColumn, MofNCompleteColumn, Progress, TextColumn,
-            TimeElapsedColumn, TimeRemainingColumn,
-        )
-        progress = Progress(
-            TextColumn("{task.description}"),
-            BarColumn(),
-            MofNCompleteColumn(),
-            TextColumn("loss={task.fields[loss]:.4f}"),
-            TextColumn("lr={task.fields[lr]:.2e}"),
-            TextColumn("speed={task.fields[speed]:.2f} it/s"),
-            TimeElapsedColumn(),
-            TimeRemainingColumn(),
-            refresh_per_second=10,
-        )
-        task = progress.add_task("train", total=total_steps, loss=0.0, lr=0.0, speed=0.0)
-        return progress, task, "rich"
-    except Exception:
-        return "plain", None, None
-
-
-def render_loss_curve(losses, width=60, height=10):
-    """渲染 ASCII Loss 曲线"""
-    if not losses:
-        return ""
-    if width < 5:
-        width = 5
-    values = losses
-    if len(values) > width:
-        step = len(values) / width
-        buckets = []
-        for i in range(width):
-            start = int(i * step)
-            end = int((i + 1) * step)
-            end = max(end, start + 1)
-            chunk = values[start:end]
-            buckets.append(sum(chunk) / len(chunk))
-        values = buckets
-    min_v = min(values)
-    max_v = max(values)
-    if max_v == min_v:
-        max_v = min_v + 1e-8
-    grid = [[" " for _ in range(len(values))] for _ in range(height)]
-    for i, v in enumerate(values):
-        y = int((v - min_v) / (max_v - min_v) * (height - 1))
-        y = height - 1 - y
-        grid[y][i] = "*"
-    lines = ["".join(row) for row in grid]
-    lines.append(f"min={min_v:.4f} max={max_v:.4f}")
-    return "\n".join(lines)
-
-
-def render_curve_panel(losses, width=60, height=10):
-    """渲染 Rich Panel 包装的 Loss 曲线"""
-    try:
-        from rich.panel import Panel
-        from rich.text import Text
-    except Exception:
-        return None
-    chart = render_loss_curve(losses, width=width, height=height)
-    return Panel(Text(chart), title="Loss curve (recent)", expand=False)
-
-
-# ============================================================================
-# Weights & Biases 可选监控
-# ============================================================================
-
-class WandBMonitor:
-    def __init__(
-        self,
-        wandb_module,
-        run,
-        *,
-        log_samples: bool = False,
-        sample_max_side: int = 512,
-        sample_every_n_steps: int = 0,
-    ) -> None:
-        self._wandb = wandb_module
-        self._run = run
-        self.log_samples = log_samples
-        self.sample_max_side = max(64, int(sample_max_side or 512))
-        self.sample_every_n_steps = max(0, int(sample_every_n_steps or 0))
-        self._last_logged_step: Optional[int] = None
-
-    @property
-    def enabled(self) -> bool:
-        return self._run is not None
-
-    def log(self, data: dict, *, step: Optional[int] = None) -> None:
-        if not self.enabled:
-            return
-        try:
-            self._run.log(data, step=step)
-        except Exception as exc:
-            logger.warning(f"W&B log 失败: {exc}")
-
-    def _should_log_step(self, key: str, step: Optional[int]) -> bool:
-        # baseline / epoch 边界一律放行；step 模式按 sample_every_n_steps 节流。
-        if self.sample_every_n_steps <= 0:
-            return True
-        if not key.startswith("samples/step"):
-            return True
-        if step is None or step <= 0:
-            return True
-        if step == self._last_logged_step:
-            return True  # 同步重复调用允许
-        return step % self.sample_every_n_steps == 0
-
-    def _prepare_image(self, image_path: Path, caption: str):
-        # 原图常 2K+，wandb 面板浏览 512px 已足够；JPEG 流量比 PNG 小一个数量级。
-        try:
-            from PIL import Image
-        except Exception:
-            return self._wandb.Image(str(image_path), caption=caption)
-        try:
-            with Image.open(image_path) as img:
-                img = img.convert("RGB")
-                max_side = self.sample_max_side
-                w, h = img.size
-                if max(w, h) > max_side:
-                    scale = max_side / float(max(w, h))
-                    new_size = (max(1, int(w * scale)), max(1, int(h * scale)))
-                    img = img.resize(new_size, Image.LANCZOS)
-                return self._wandb.Image(img, caption=caption)
-        except Exception as exc:
-            logger.warning(f"W&B 图片缩放失败，改传原图: {exc}")
-            return self._wandb.Image(str(image_path), caption=caption)
-
-    def log_image(self, key: str, image_path: Path, *, caption: str, step: Optional[int] = None) -> None:
-        if not self.enabled:
-            return
-        if not self._should_log_step(key, step):
-            return
-        try:
-            self._run.log({key: [self._prepare_image(image_path, caption)]}, step=step)
-            self._last_logged_step = step
-        except Exception as exc:
-            logger.warning(f"W&B 图片记录失败: {exc}")
-
-    def finish(self) -> None:
-        if not self.enabled:
-            return
-        try:
-            self._run.finish()
-        except Exception as exc:
-            logger.warning(f"W&B finish 失败: {exc}")
-
-
-def init_wandb_monitor(args, output_dir: Path, config_path: Optional[Path]) -> WandBMonitor:
-    enabled = str(os.environ.get("WANDB_ENABLED", "")).strip().lower()
-    if enabled not in {"1", "true", "yes", "on"}:
-        return WandBMonitor(None, None)
-    mode = str(os.environ.get("WANDB_MODE", "online") or "online")
-    if mode == "disabled":
-        return WandBMonitor(None, None)
-    try:
-        import wandb
-    except ImportError as exc:
-        raise RuntimeError(
-            "已在 Settings 启用 WandB，但当前环境没有安装 wandb。"
-            "请先在训练环境安装：pip install wandb，或在 Settings 关闭 WandB。"
-        ) from exc
-
-    project = os.environ.get("WANDB_PROJECT") or "AnimaLoraStudio"
-    entity = os.environ.get("WANDB_ENTITY") or None
-    run_name = os.environ.get("WANDB_RUN_NAME") or str(args.output_name)
-    # 默认关 — supervisor 已经按 secrets.wandb.log_samples 设过 env；env 缺省时也保持关闭。
-    log_samples = str(os.environ.get("WANDB_LOG_SAMPLES", "0")).strip().lower() not in {
-        "0", "false", "no", "off",
-    }
-    try:
-        sample_max_side = int(os.environ.get("WANDB_SAMPLE_MAX_SIDE", "512") or 512)
-    except ValueError:
-        sample_max_side = 512
-    try:
-        sample_every_n_steps = int(os.environ.get("WANDB_SAMPLE_EVERY_N_STEPS", "0") or 0)
-    except ValueError:
-        sample_every_n_steps = 0
-    wandb_dir = output_dir / "wandb"
-    wandb_dir.mkdir(parents=True, exist_ok=True)
-    cfg = {
-        key: value
-        for key, value in vars(args).items()
-        if key not in {"interactive", "auto_install"}
-    }
-    cfg["config_path"] = str(config_path) if config_path else ""
-    run = wandb.init(
-        project=project,
-        entity=entity,
-        name=run_name,
-        mode=mode,
-        config=cfg,
-        dir=str(wandb_dir),
-    )
-    logger.info(f"W&B 监控已启用: project={project}, run={run_name}, mode={mode}")
-    return WandBMonitor(
-        wandb,
-        run,
-        log_samples=log_samples,
-        sample_max_side=sample_max_side,
-        sample_every_n_steps=sample_every_n_steps,
-    )
-
-
-# ============================================================================
-# 梯度检查点
-# ============================================================================
-
-def forward_with_optional_checkpoint(model, latents, timesteps, cross, padding_mask, use_checkpoint=False):
-    """带可选梯度检查点的前向传播"""
-    if not use_checkpoint:
-        return model(latents, timesteps, cross, padding_mask=padding_mask)
-    from torch.utils.checkpoint import checkpoint
-
-    x_B_T_H_W_D, rope_emb, extra_pos_emb = model.prepare_embedded_sequence(
-        latents, fps=None, padding_mask=padding_mask,
-    )
-    if timesteps.ndim == 1:
-        timesteps = timesteps.unsqueeze(1)
-    t_embedding, adaln_lora = model.t_embedder(timesteps)
-    t_embedding = model.t_embedding_norm(t_embedding)
-
-    block_kwargs = {
-        "rope_emb_L_1_1_D": rope_emb,
-        "adaln_lora_B_T_3D": adaln_lora,
-        "extra_per_block_pos_emb": extra_pos_emb,
-    }
-
-    for block in model.blocks:
-        def custom_forward(x, blk=block):
-            return blk(x, t_embedding, cross, **block_kwargs)
-        x_B_T_H_W_D = checkpoint(custom_forward, x_B_T_H_W_D, use_reentrant=False)
-
-    x_B_T_H_W_O = model.final_layer(x_B_T_H_W_D, t_embedding, adaln_lora_B_T_3D=adaln_lora)
-    return model.unpatchify(x_B_T_H_W_O)
-
-
-# ============================================================================
-# xformers 支持
-# ============================================================================
-
-def enable_xformers(model):
-    """为模型启用 xformers memory efficient attention"""
-    try:
-        from xformers.ops import memory_efficient_attention
-    except ImportError:
-        logger.warning("xformers 未安装，跳过启用")
-        return False
-
-    enabled_count = 0
-    for name, module in model.named_modules():
-        # 查找 attention 模块并替换
-        if hasattr(module, "set_use_memory_efficient_attention_xformers"):
-            module.set_use_memory_efficient_attention_xformers(True)
-            enabled_count += 1
-        elif hasattr(module, "enable_xformers_memory_efficient_attention"):
-            module.enable_xformers_memory_efficient_attention()
-            enabled_count += 1
-
-    if enabled_count > 0:
-        logger.info(f"xformers 已启用: {enabled_count} 个模块")
-        return True
-
-    # 如果模型没有内置支持，尝试 monkey patch
-    logger.info("xformers 已加载，将在 attention 计算中使用")
-    return True
-
-
-# ============================================================================
-# 模型加载工具
-# ============================================================================
-
-def find_diffusion_pipe_root():
-    """查找 diffusion-pipe 模型代码路径。
-
-    候选顺序（首个命中即返回）：
-      1. 脚本同目录 `diffusion_models/` / `models/`（CLI 直接 cd 进 scripts/ 跑）
-      2. 仓库根 `models/` / `diffusion_models/`（训练脚本在 runtime/ 下的 repo
-         layout：repo_root/runtime/anima_train.py → ../models/anima_modeling.py）
-      3. 环境变量 `DIFFUSION_PIPE_ROOT`（覆盖路径用）
-    """
-    script_dir = Path(__file__).parent
-    repo_root = script_dir.parent
-    candidates = [
-        script_dir / "diffusion_models",
-        script_dir / "models",
-        repo_root / "models",
-        repo_root / "diffusion_models",
-        Path(os.environ.get("DIFFUSION_PIPE_ROOT", "")) if os.environ.get("DIFFUSION_PIPE_ROOT") else None,
-    ]
-    for candidate in candidates:
-        if candidate and (candidate / "anima_modeling.py").exists():
-            return candidate
-        if candidate and (candidate / "models" / "anima_modeling.py").exists():
-            return candidate / "models"
-    raise RuntimeError("找不到 anima_modeling.py，请设置 DIFFUSION_PIPE_ROOT 或放置模型代码")
-
-
-def load_module_from_path(module_name, file_path):
-    """动态加载 Python 模块"""
-    import importlib.util
-    spec = importlib.util.spec_from_file_location(module_name, file_path)
-    module = importlib.util.module_from_spec(spec)
-    spec.loader.exec_module(module)
-    return module
-
-
-def _strip_prefixes(key: str, prefixes: list[str]) -> str:
-    """反复剥离前缀（支持 module.model. 这种复合前缀）"""
-    if not prefixes:
-        return key
-    changed = True
-    while changed:
-        changed = False
-        for p in prefixes:
-            if key.startswith(p):
-                key = key[len(p) :]
-                changed = True
-    return key
-
-
-def _pick_best_prefix_remap(sd_keys: list[str], model_keys: set[str]) -> tuple[list[str], int]:
-    """
-    从常见前缀组合里选择“命中最多 model_keys”的 remap 方案。
-    返回 (prefixes, matched_count)。
-    """
-    candidates: list[tuple[str, list[str]]] = [
-        ("none", []),
-        ("net.", ["net."]),
-        ("model.", ["model."]),
-        ("module.", ["module."]),
-        ("module.+model.", ["module.", "model."]),
-        ("module.model.", ["module.model."]),
-        ("diffusion_model.", ["diffusion_model."]),
-        ("model.diffusion_model.", ["model.diffusion_model."]),
-        ("transformer.", ["transformer."]),
-        ("vae.", ["vae."]),
-        ("first_stage_model.", ["first_stage_model."]),
-        ("net.+model.", ["net.", "model."]),
-        ("net.model.", ["net.model."]),
-    ]
-
-    best_prefixes: list[str] = []
-    best_matched = -1
-    for _name, prefixes in candidates:
-        matched = 0
-        for k in sd_keys:
-            kk = _strip_prefixes(k, prefixes)
-            if kk in model_keys:
-                matched += 1
-        if matched > best_matched:
-            best_matched = matched
-            best_prefixes = prefixes
-    return best_prefixes, best_matched
-
-
-def _load_safetensors_state_dict(path: Path) -> dict:
-    from safetensors import safe_open
-
-    sd = {}
-    with safe_open(path, framework="pt", device="cpu") as f:
-        for k in f.keys():
-            sd[k] = f.get_tensor(k)
-    return sd
-
-
-def resolve_path_best_effort(path_str: str, bases: list[Path]) -> str:
-    """
-    将相对路径按多个 base 尝试解析到一个真实存在的路径。
-    主要用于：无论从 repo 根目录还是 AnimaLoraToolkit 目录启动，都能找到 models/* 文件。
-    """
-    if not path_str:
-        return path_str
-
-    p = Path(path_str)
-    if p.is_absolute():
-        return str(p)
-
-    # 先按原样（相对 cwd）试一下
-    if p.exists():
-        return str(p)
-
-    # 逐 base 拼接尝试
-    for b in bases:
-        if not b:
-            continue
-        try:
-            cand = (Path(b) / p).resolve()
-        except Exception:
-            cand = Path(b) / p
-        if cand.exists():
-            return str(cand)
-
-    # 常见：配置写了 AnimaLoraToolkit/xxx，但启动目录已经在 AnimaLoraToolkit 下
-    parts = p.parts
-    if parts and parts[0].lower() in ("animaloratoolkit", "anima_trainer", "anima-trainer"):
-        p2 = Path(*parts[1:])
-        if p2.exists():
-            return str(p2)
-        for b in bases:
-            if not b:
-                continue
-            cand = Path(b) / p2
-            if cand.exists():
-                return str(cand)
-
-    return path_str
-
-
-def _load_weights_best_effort(model: torch.nn.Module, sd: dict, label: str) -> dict:
-    """
-    更健壮的权重加载：
-    - 自动尝试剥离常见前缀（model./module./...）
-    - 打印匹配率、missing/unexpected
-    - 关键模块未加载时直接报错（避免“采样全噪点”还继续训练）
-    """
-    model_keys = set(model.state_dict().keys())
-    sd_keys = list(sd.keys())
-    prefixes, matched = _pick_best_prefix_remap(sd_keys, model_keys)
-    remapped = {_strip_prefixes(k, prefixes): v for k, v in sd.items()}
-
-    incompatible = model.load_state_dict(remapped, strict=False)
-    missing = list(getattr(incompatible, "missing_keys", []) or [])
-    unexpected = list(getattr(incompatible, "unexpected_keys", []) or [])
-
-    matched_after = len(set(remapped.keys()) & model_keys)
-    coverage = matched_after / max(1, len(model_keys))
-    remap_name = "+".join(prefixes) if prefixes else "none"
-
-    logger.info(
-        f"{label} 权重加载: remap={remap_name}, 匹配 {matched_after}/{len(model_keys)} ({coverage:.1%}), "
-        f"missing={len(missing)}, unexpected={len(unexpected)}"
-    )
-
-    # 关键层缺失会直接导致输出接近 0，采样就是纯噪点
-    critical_prefixes = ("x_embedder.", "blocks.", "final_layer.")
-    critical_missing = [k for k in missing if k.startswith(critical_prefixes)]
-    if coverage < 0.60 or len(critical_missing) > 0:
-        preview_missing = ", ".join(critical_missing[:8])
-        raise RuntimeError(
-            f"{label} 权重看起来没有正确加载（remap={remap_name}, coverage={coverage:.1%}）。"
-            f"关键参数缺失: {preview_missing or 'N/A'}。\n"
-            f"这通常表示你选错了 .safetensors（不是完整 transformer/vae 权重），或 checkpoint key 前缀不匹配。"
-        )
-    return {
-        "remap": remap_name,
-        "coverage": coverage,
-        "missing": missing,
-        "unexpected": unexpected,
-    }
+# ADR 0003 PR-A：模块化拆分。下面这一段从 training.* re-export，给 sister script /
+# test 维持 `anima_train.X` 的访问路径。新代码请直接 import 子模块。
+from training.bootstrap import (  # noqa: E402
+    apply_yaml_config,
+    ensure_dependencies,
+    init_progress,
+    load_yaml_config,
+)
+from training.observability import (  # noqa: E402
+    WandBMonitor,
+    init_wandb_monitor,
+    render_curve_panel,
+    render_loss_curve,
+)
+from training.model_loading import (  # noqa: E402
+    _load_safetensors_state_dict,
+    _load_weights_best_effort,
+    _pick_best_prefix_remap,
+    _strip_prefixes,
+    enable_xformers,
+    find_diffusion_pipe_root,
+    forward_with_optional_checkpoint,
+    load_module_from_path,
+    resolve_path_best_effort,
+)
+from training.text_encoding import (  # noqa: E402
+    _build_qwen_text_from_prompt,
+    _parse_weighted_tag,
+    encode_qwen,
+    tokenize_t5_weighted,
+)
+from training.state import load_training_state, save_training_state  # noqa: E402
 
 
 def ensure_models_namespace(repo_root):
@@ -770,53 +246,9 @@ def load_text_encoders(qwen_path, t5_tokenizer_path, device, dtype):
     return qwen_model, qwen_tokenizer, t5_tokenizer
 
 
-# ============================================================================
-# 文本编码
-# ============================================================================
-
-def encode_qwen(model, tokenizer, texts, device, max_length=512):
-    """Qwen 文本编码"""
-    # Qwen3 tokenizer 对空字符串可能返回 0 tokens（会导致模型内部 reshape 失败）
-    # ComfyUI 的 AnimaTokenizer 设置了 min_length=1，这里做同等兜底。
-    if isinstance(texts, str):
-        texts = [texts]
-    texts = [(" " if (t is None or str(t).strip() == "") else str(t)) for t in texts]
-
-    inputs = tokenizer(
-        texts,
-        return_tensors="pt",
-        padding=True,
-        truncation=True,
-        max_length=max_length,
-        add_special_tokens=False,
-    )
-    # 仍可能出现空序列（极端 tokenizer 行为），强制塞 1 个 token
-    if inputs["input_ids"].ndim == 2 and inputs["input_ids"].shape[1] == 0:
-        pad_id = tokenizer.pad_token_id
-        if pad_id is None:
-            pad_id = tokenizer.eos_token_id
-        if pad_id is None:
-            pad_id = 0
-        bs = len(texts)
-        inputs["input_ids"] = torch.full((bs, 1), int(pad_id), dtype=torch.long)
-        inputs["attention_mask"] = torch.ones((bs, 1), dtype=torch.long)
-    inputs = inputs.to(device)
-
-    with torch.inference_mode():
-        outputs = model(
-            input_ids=inputs["input_ids"],
-            attention_mask=inputs["attention_mask"],
-            output_hidden_states=True,
-            return_dict=True,
-            use_cache=False,
-        )
-
-    hidden = outputs.hidden_states[-1]
-    # 清零 padding 位置
-    mask = inputs["attention_mask"].unsqueeze(-1)
-    hidden = hidden * mask
-
-    return hidden, inputs["attention_mask"]
+# 文本编码相关（encode_qwen / _parse_weighted_tag / _build_qwen_text_from_prompt /
+# tokenize_t5_weighted）已搬到 training.text_encoding（ADR 0003 PR-A）。
+# 顶部 import 已 re-export，sister script / test 不需改。
 
 
 def _time_snr_shift(alpha: float, t: torch.Tensor) -> torch.Tensor:
@@ -960,181 +392,11 @@ def _sample_er_sde_const_x0(
     return x
 
 
-def _parse_weighted_tag(tag: str) -> tuple[str, float]:
-    """
-    解析单个 tag 的权重（参考指南“权重控制”）。
-    支持：
-    - (tag:1.5)
-    - (tag) / ((tag))  => 1.1^n
-    - [tag]            => 1/1.1
-    """
-    import re
-
-    s = tag.strip()
-    if not s:
-        return "", 1.0
-
-    # 显式 (xxx:1.23)
-    m = re.fullmatch(r"\(\s*(.+?)\s*:\s*([+-]?\d+(?:\.\d+)?)\s*\)", s)
-    if m:
-        return m.group(1).strip(), float(m.group(2))
-
-    # 统计外层 () / [] 深度
-    w = 1.0
-    while True:
-        s2 = s.strip()
-        if len(s2) >= 2 and s2[0] == "(" and s2[-1] == ")":
-            s = s2[1:-1].strip()
-            w *= 1.1
-            continue
-        if len(s2) >= 2 and s2[0] == "[" and s2[-1] == "]":
-            s = s2[1:-1].strip()
-            w /= 1.1
-            continue
-        break
-    return s.strip(), float(w)
-
-
-def _build_qwen_text_from_prompt(prompt: str) -> str:
-    # Qwen 通道不传权重，只传“干净标签文本”（参考 ComfyUI anima-kai 的做法）
-    parts = [p.strip() for p in prompt.split(",") if p.strip()]
-    clean = []
-    for p in parts:
-        t, _w = _parse_weighted_tag(p)
-        if t:
-            clean.append(t)
-    return ", ".join(clean)
-
-
-def tokenize_t5_weighted(tokenizer, texts, max_length=512):
-    """
-    参考 ComfyUI 的 anima-kai：按逗号切分 tag，逐 tag 分词，并为每个 token 附带权重。
-    返回：input_ids, attention_mask(1=有效), token_weights
-    """
-    import torch
-
-    if isinstance(texts, str):
-        texts = [texts]
-
-    pad_id = tokenizer.pad_token_id if tokenizer.pad_token_id is not None else 0
-    eos_id = tokenizer.eos_token_id if tokenizer.eos_token_id is not None else 1
-
-    all_ids = []
-    all_w = []
-    for text in texts:
-        tags = [t.strip() for t in str(text).split(",") if t.strip()]
-        ids = []
-        ws = []
-        for tag in tags:
-            clean_tag, weight = _parse_weighted_tag(tag)
-            if not clean_tag:
-                continue
-            tok = tokenizer(clean_tag, add_special_tokens=False)
-            for tid in tok["input_ids"]:
-                ids.append(int(tid))
-                ws.append(float(weight))
-
-        # 末尾补一个 eos（ComfyUI 也是最后加一个终止 token）
-        ids.append(int(eos_id))
-        ws.append(1.0)
-
-        # 截断到 max_length（保留最后一个 eos）
-        if max_length and len(ids) > max_length:
-            ids = ids[: max_length - 1] + [int(eos_id)]
-            ws = ws[: max_length - 1] + [1.0]
-
-        all_ids.append(torch.tensor(ids, dtype=torch.long))
-        all_w.append(torch.tensor(ws, dtype=torch.float32))
-
-    # pad 到 batch 内最长
-    max_len = max(x.numel() for x in all_ids) if all_ids else 1
-    input_ids = torch.full((len(all_ids), max_len), pad_id, dtype=torch.long)
-    token_w = torch.zeros((len(all_w), max_len), dtype=torch.float32)
-    attention_mask = torch.zeros((len(all_ids), max_len), dtype=torch.long)
-
-    for i, (ids, ws) in enumerate(zip(all_ids, all_w)):
-        L = ids.numel()
-        input_ids[i, :L] = ids
-        token_w[i, :L] = ws
-        attention_mask[i, :L] = 1
-
-    return input_ids, attention_mask, token_w
-
-
-# ============================================================================
 # LoRA / LoKr 实现：见 utils.lycoris_adapter.AnimaLycorisAdapter
 # （历史自实现版本于 Stage 3c 删除，使用 lycoris-lora 包替代）
-# ============================================================================
 
-# ============================================================================
-# 训练状态保存/恢复（断点续训）
-# ============================================================================
-
-def save_training_state(path, injector, optimizer, epoch, global_step, loss_history=None, rng_state=None, monitor_state=None, scheduler=None):
-    """保存完整训练状态，支持断点续训"""
-    state = {
-        "lora_state_dict": injector.state_dict(),
-        "optimizer_state_dict": optimizer.state_dict(),
-        "epoch": epoch,
-        "global_step": global_step,
-        "loss_history": loss_history or [],
-        "rng_state": {
-            "torch": torch.get_rng_state(),
-            "cuda": torch.cuda.get_rng_state() if torch.cuda.is_available() else None,
-            "random": random.getstate(),
-        },
-        "monitor_state": monitor_state,  # 保存监控面板数据（用于恢复 loss 曲线）
-    }
-    if scheduler is not None:
-        state["scheduler_state_dict"] = scheduler.state_dict()
-    torch.save(state, path)
-    logger.info(f"训练状态已保存: {path} (epoch={epoch}, step={global_step})")
-
-
-def load_training_state(path, injector, optimizer, scheduler=None):
-    """加载训练状态，返回 (epoch, global_step, loss_history, monitor_state)"""
-    logger.info(f"加载训练状态: {path}")
-    state = torch.load(path, map_location="cpu", weights_only=False)
-    
-    # 加载 LoRA 权重（lycoris-lora backend）— 一次性导入 state_dict
-    # 旧自实现 ckpt 在 Stage 4 plan 决策中**不做迁移**，strict=False 让缺失键
-    # 走默认初始化路径而非崩溃；用户应当从头训练新格式 ckpt。
-    lora_sd = state["lora_state_dict"]
-    result = injector.load_state_dict(lora_sd, strict=False)
-    missing = len(getattr(result, "missing_keys", [])) if hasattr(result, "missing_keys") else 0
-    unexpected = len(getattr(result, "unexpected_keys", [])) if hasattr(result, "unexpected_keys") else 0
-    if missing or unexpected:
-        logger.warning(
-            f"resume LoRA: missing={missing}, unexpected={unexpected}（旧格式 ckpt？）"
-        )
-    
-    # 加载优化器状态
-    optimizer.load_state_dict(state["optimizer_state_dict"])
-
-    # 加载调度器状态
-    if scheduler is not None and "scheduler_state_dict" in state:
-        try:
-            scheduler.load_state_dict(state["scheduler_state_dict"])
-        except Exception as e:
-            logger.warning(f"调度器状态恢复失败（将从头开始）: {e}")
-
-    # 恢复随机数状态
-    if "rng_state" in state:
-        rng = state["rng_state"]
-        if rng.get("torch") is not None:
-            torch.set_rng_state(rng["torch"])
-        if rng.get("cuda") is not None and torch.cuda.is_available():
-            torch.cuda.set_rng_state(rng["cuda"])
-        if rng.get("random") is not None:
-            random.setstate(rng["random"])
-    
-    epoch = state.get("epoch", 0)
-    global_step = state.get("global_step", 0)
-    loss_history = state.get("loss_history", [])
-    monitor_state = state.get("monitor_state", None)  # 恢复监控数据
-    
-    logger.info(f"训练状态已恢复: epoch={epoch}, step={global_step}")
-    return epoch, global_step, loss_history, monitor_state
+# 训练状态保存/恢复（save_training_state / load_training_state）已搬到
+# training.state；caption / tokenize 已搬到 training.text_encoding（ADR 0003 PR-A）。
 
 
 # ============================================================================
